--- tests/test_transforms.py
@@ -1,985 +1,0 @@
-from __future__ import annotations
-
-from datetime import datetime, timezone
-from math import isclose
-from typing import Any
-
-import pytest
-from datapipeline.domain.feature import FeatureRecord
-from datapipeline.domain.record import TemporalRecord
-from datapipeline.domain.sample import Sample
-from datapipeline.domain.vector import Vector
-from datapipeline.transforms.feature.scaler import StandardScaler, StandardScalerTransform
-from datapipeline.transforms.stream.dedupe import FeatureDeduplicateTransform
-from datapipeline.transforms.stream.fill import FillTransformer as FeatureFill
-from datapipeline.transforms.vector import (
-    VectorDropTransform,
-    VectorEnsureSchemaTransform,
-    VectorReplaceTransform,
-    VectorFillTransform,
-)
-
-
-def _make_time_record(value: float, hour: int) -> TemporalRecord:
-    return TemporalRecord(
-        time=datetime(2024, 1, 1, hour=hour, tzinfo=timezone.utc),
-        value=value,
-    )
-
-
-def _make_feature_record(value: float, hour: int, feature_id: str) -> FeatureRecord:
-    return FeatureRecord(
-        record=_make_time_record(value, hour),
-        id=feature_id,
-    )
-
-
-def _make_vector(group: int, values: dict[str, Any]) -> Sample:
-    return Sample(key=(group,), features=Vector(values=values))
-
-
-class _StubVectorContext:
-    def __init__(
-        self,
-        expected: list[str] | dict[str, list[str]],
-        *,
-        schema: dict[str, list[dict]] | None = None,
-    ):
-        if isinstance(expected, dict):
-            self._expected_map = {k: list(v) for k, v in expected.items()}
-        else:
-            self._expected_map = {"features": list(expected)}
-        self._schema_map = schema or {}
-        self._cache: dict[str, Any] = {}
-        self._metadata_doc: dict | None = None
-
-    def load_expected_ids(self, *, payload: str = "features") -> list[str]:
-        ids = list(self._expected_map.get(payload, []))
-        self._cache[f"expected_ids:{payload}"] = list(ids)
-        return ids
-
-    def load_schema(self, *, payload: str = "features") -> list[dict]:
-        entries = self._schema_map.get(payload)
-        if entries is not None:
-            snapshot = [dict(item) for item in entries]
-        else:
-            expected = self._expected_map.get(payload)
-            if expected is None:
-                snapshot = []
-            else:
-                snapshot = [{"id": fid} for fid in expected]
-        self._cache[f"schema:{payload}"] = [dict(item) for item in snapshot]
-        self._cache[f"expected_ids:{payload}"] = [
-            entry["id"] for entry in snapshot if isinstance(entry.get("id"), str)
-        ]
-        return snapshot
-
-    def set_metadata(self, doc: dict) -> None:
-        self._metadata_doc = doc
-
-    def require_artifact(self, spec) -> dict:
-        if self._metadata_doc is None:
-            raise RuntimeError("artifact missing")
-        return self._metadata_doc
-
-
-def test_standard_scaler_normalizes_feature_stream():
-    training_vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"radiation": 1.0})),
-            Sample(key=(1,), features=Vector(values={"radiation": 2.0})),
-            Sample(key=(2,), features=Vector(values={"radiation": 3.0})),
-        ]
-    )
-    scaler_model = StandardScaler()
-    scaler_model.fit(training_vectors)
-    scaler = StandardScalerTransform(scaler=scaler_model)
-
-    stream = iter(
-        [
-            _make_feature_record(1.0, 0, "radiation"),
-            _make_feature_record(2.0, 1, "radiation"),
-            _make_feature_record(3.0, 2, "radiation"),
-        ]
-    )
-
-    transformed = list(scaler.apply(stream))
-
-    values = [fr.record.value for fr in transformed]
-    expected = [-1.22474487, 0.0, 1.22474487]
-    for observed, target in zip(values, expected):
-        assert isclose(observed, target, rel_tol=1e-6)
-
-
-def test_standard_scaler_uses_provided_statistics():
-    training_vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"temperature": 0.0})),
-            Sample(key=(1,), features=Vector(values={"temperature": 10.0})),
-        ]
-    )
-    scaler_model = StandardScaler()
-    scaler_model.fit(training_vectors)
-    scaler = StandardScalerTransform(scaler=scaler_model)
-
-    stream = iter(
-        [
-            _make_feature_record(10.0, 0, "temperature"),
-            _make_feature_record(11.0, 1, "temperature"),
-        ]
-    )
-
-    transformed = list(scaler.apply(stream))
-
-    assert [fr.record.value for fr in transformed] == [1.0, 1.2]
-
-
-def test_standard_scaler_inverse_transform_round_trip():
-    training_vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"radiation": 1.0})),
-            Sample(key=(1,), features=Vector(values={"radiation": 2.0})),
-            Sample(key=(2,), features=Vector(values={"radiation": 3.0})),
-        ]
-    )
-    scaler_model = StandardScaler()
-    scaler_model.fit(training_vectors)
-    transform = StandardScalerTransform(scaler=scaler_model)
-
-    stream = iter(
-        [
-            _make_feature_record(1.0, 0, "radiation"),
-            _make_feature_record(2.0, 1, "radiation"),
-            _make_feature_record(3.0, 2, "radiation"),
-        ]
-    )
-
-    scaled = list(transform.apply(stream))
-    restored = list(transform.inverse(iter(scaled)))
-
-    original = [1.0, 2.0, 3.0]
-    values = [fr.record.value for fr in restored]
-    for observed, expected in zip(values, original):
-        assert isclose(observed, expected, rel_tol=1e-6)
-
-
-def test_standard_scaler_fit_and_serialize(tmp_path):
-    vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"temp": 10.0, "wind": 5.0})),
-            Sample(key=(1,), features=Vector(values={"temp": 14.0, "wind": 7.0})),
-        ]
-    )
-
-    scaler = StandardScaler()
-    total = scaler.fit(vectors)
-    assert total == 4
-    path = tmp_path / "scaler.pkl"
-    scaler.save(path)
-    restored = StandardScaler.load(path)
-    transform = StandardScalerTransform(model_path=path)
-
-    stream = iter(
-        [
-            _make_feature_record(12.0, 0, "temp"),
-            _make_feature_record(6.0, 0, "wind"),
-        ]
-    )
-
-    transformed = list(transform.apply(stream))
-    assert len(restored.statistics) == 2
-    assert len(transformed) == 2
-
-
-def test_standard_scaler_errors_on_missing_by_default():
-    training_vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"temp": 10.0})),
-            Sample(key=(1,), features=Vector(values={"temp": 14.0})),
-        ]
-    )
-    scaler_model = StandardScaler()
-    scaler_model.fit(training_vectors)
-    transform = StandardScalerTransform(scaler=scaler_model, on_none="error")
-
-    stream = iter(
-        [
-            _make_feature_record(10.0, 0, "temp"),
-            _make_feature_record(None, 1, "temp"),
-        ]
-    )
-
-    with pytest.raises(TypeError):
-        list(transform.apply(stream))
-
-
-def test_standard_scaler_passthrough_missing_counts():
-    training_vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"temp": 1.0})),
-            Sample(key=(1,), features=Vector(values={"temp": 3.0})),
-        ]
-    )
-    scaler_model = StandardScaler()
-    scaler_model.fit(training_vectors)
-
-    transform = StandardScalerTransform(scaler=scaler_model, on_none="skip")
-
-    stream = iter(
-        [
-            _make_feature_record(1.0, 0, "temp"),
-            _make_feature_record(None, 1, "temp"),
-            _make_feature_record(3.0, 2, "temp"),
-        ]
-    )
-
-    transformed = list(transform.apply(stream))
-    values = [fr.record.value for fr in transformed]
-    assert values == [-1.0, None, 1.0]
-    assert transform.missing_counts == {"temp": 1}
-
-
-def test_standard_scaler_matches_sklearn():
-    sklearn = pytest.importorskip("sklearn.preprocessing")
-    SKStandardScaler = getattr(sklearn, "StandardScaler")
-
-    values = [1.0, 2.0, 3.0, 4.0]
-    vectors = iter(
-        [Sample(key=(i,), features=Vector(values={"x": v})) for i, v in enumerate(values)]
-    )
-
-    scaler = StandardScaler()
-    scaler.fit(vectors)
-
-    sk_scaler = SKStandardScaler()
-    sk_scaler.fit([[v] for v in values])
-
-    stream = iter([_make_feature_record(v, i, "x") for i, v in enumerate(values)])
-    transformed = list(StandardScalerTransform(scaler=scaler).apply(stream))
-    ours = [fr.record.value for fr in transformed]
-    theirs = sk_scaler.transform([[v] for v in values]).flatten().tolist()
-
-    assert pytest.approx(ours) == theirs
-
-
-def test_standard_scaler_warn_callback_invoked_with_counts():
-    training_vectors = iter(
-        [
-            Sample(key=(0,), features=Vector(values={"temp": 1.0})),
-            Sample(key=(1,), features=Vector(values={"temp": 3.0})),
-        ]
-    )
-    scaler_model = StandardScaler()
-    scaler_model.fit(training_vectors)
-
-    calls: list[tuple[str, float | None, int]] = []
-
-    def on_none_cb(event):
-        if event.type == "scaler_none":
-            calls.append(
-                (
-                    event.payload["feature_id"],
-                    event.payload["record"].time.hour,
-                    event.payload["count"],
-                )
-            )
-        elif event.type == "scaler_none_summary":
-            calls.append(("summary", None, event.payload["count"]))
-
-    transform = StandardScalerTransform(scaler=scaler_model, on_none="skip")
-    transform.set_observer(on_none_cb)
-
-    stream = iter(
-        [
-            _make_feature_record(1.0, 0, "temp"),
-            _make_feature_record(None, 1, "temp"),
-            _make_feature_record(None, 2, "temp"),
-        ]
-    )
-
-    transformed = list(transform.apply(stream))
-
-    assert [fr.record.value for fr in transformed] == [-1.0, None, None]
-    assert transform.missing_counts == {"temp": 2}
-    assert calls == [("temp", 1, 1), ("temp", 2, 2)]
-
-
-def test_time_mean_fill_uses_running_average():
-    stream = iter(
-        [
-            _make_feature_record(10.0, 0, "temp"),
-            _make_feature_record(12.0, 1, "temp"),
-            _make_feature_record(None, 2, "temp"),
-            _make_feature_record(16.0, 3, "temp"),
-            _make_feature_record(float("nan"), 4, "temp"),
-        ]
-    )
-
-    transformer = FeatureFill(statistic="mean", window=2)
-
-    transformed = list(transformer.apply(stream))
-    values = [fr.record.value for fr in transformed]
-
-    assert values[2] == 11.0  # mean of 10 and 12
-    assert values[4] == 16.0  # window counts ticks, so only the fresh valid value is available
-
-
-def test_time_median_fill_honours_window():
-    stream = iter(
-        [
-            _make_feature_record(1.0, 0, "wind"),
-            _make_feature_record(100.0, 1, "wind"),
-            _make_feature_record(2.0, 2, "wind"),
-            _make_feature_record(None, 3, "wind"),
-            _make_feature_record(None, 4, "wind"),
-        ]
-    )
-
-    transformer = FeatureFill(statistic="median", window=2)
-
-    transformed = list(transformer.apply(stream))
-    values = [fr.record.value for fr in transformed]
-
-    # history window restricted to last two valid values -> [100, 2]
-    assert values[3] == 51.0
-    # second missing only sees one valid tick (2.0) because window tracks ticks, not values
-    assert values[4] == 2.0
-
-
-def test_stream_dedupe_removes_exact_duplicates():
-    stream = iter(
-        [
-            _make_feature_record(10.0, 0, "temp"),
-            _make_feature_record(10.0, 0, "temp"),
-            _make_feature_record(12.0, 1, "temp"),
-            _make_feature_record(5.0, 0, "wind"),
-            _make_feature_record(5.0, 0, "wind"),
-        ]
-    )
-    transform = FeatureDeduplicateTransform()
-    out = list(transform.apply(stream))
-    assert [fr.record.value for fr in out] == [10.0, 12.0, 5.0]
-
-
-def test_stream_dedupe_keeps_distinct_values():
-    stream = iter(
-        [
-            _make_feature_record(10.0, 0, "temp"),
-            _make_feature_record(11.0, 0, "temp"),
-        ]
-    )
-    transform = FeatureDeduplicateTransform()
-    out = list(transform.apply(stream))
-    assert [fr.record.value for fr in out] == [10.0, 11.0]
-
-
-def test_vector_fill_history_uses_running_statistics():
-    stream = iter(
-        [
-            _make_vector(0, {"temp__A": 10.0}),
-            _make_vector(1, {"temp__A": 12.0}),
-            _make_vector(2, {}),
-        ]
-    )
-
-    transform = VectorFillTransform(
-        statistic="mean", window=2, min_samples=2)
-    transform.bind_context(_StubVectorContext(["temp__A"]))
-
-    out = list(transform.apply(stream))
-    assert out[2].features.values["temp__A"] == 11.0
-
-
-def test_vector_fill_constant_injects_value():
-    stream = iter([_make_vector(0, {"time": 1.0})])
-    transform = VectorReplaceTransform(value=0.0)
-    transform.bind_context(_StubVectorContext(["time", "wind"]))
-    out = list(transform.apply(stream))
-    assert out[0].features.values["wind"] == 0.0
-
-
-def test_vector_drop_horizontal_respects_coverage_with_schema():
-    stream = iter(
-        [
-            _make_vector(0, {"a": 1.0, "b": 2.0}),
-            _make_vector(1, {"a": 3.0}),
-        ]
-    )
-    transform = VectorDropTransform(axis="horizontal", threshold=1.0)
-    transform.bind_context(_StubVectorContext(["a", "b"]))
-
-    out = list(transform.apply(stream))
-    assert len(out) == 1
-    assert out[0].features.values == {"a": 1.0, "b": 2.0}
-
-
-def test_vector_fill_constant_targets_payload():
-    sample = Sample(
-        key=(0,),
-        features=Vector(values={"f": 1.0}),
-        targets=Vector(values={"t": None}),
-    )
-    transform = VectorReplaceTransform(value=5.0, payload="targets")
-    transform.bind_context(_StubVectorContext({"targets": ["t"]}))
-
-    out = list(transform.apply(iter([sample])))
-
-    assert out[0].features.values == {"f": 1.0}
-    assert out[0].targets is not None
-    assert out[0].targets.values["t"] == 5.0
-
-
-def test_vector_drop_horizontal_drops_sparse_vector():
-    stream = iter(
-        [
-            _make_vector(0, {"a": 1.0, "b": 2.0}),
-            _make_vector(1, {"a": 3.0, "b": None}),
-        ]
-    )
-    transform = VectorDropTransform(axis="horizontal", threshold=1.0)
-    transform.bind_context(_StubVectorContext(["a", "b"]))
-
-    out = list(transform.apply(stream))
-    assert len(out) == 1
-    assert out[0].features.values == {"a": 1.0, "b": 2.0}
-
-
-def test_vector_drop_horizontal_sequence_aware():
-    stream = iter(
-        [
-            _make_vector(0, {"seq": [1.0, None, None], "f": 2.0}),
-        ]
-    )
-    # seq coverage = 1/3, f coverage = 1.0 -> horizontal = (1/3 + 1)/2 ≈ 0.666
-    transform = VectorDropTransform(axis="horizontal", threshold=0.7)
-    transform.bind_context(_StubVectorContext(["seq", "f"]))
-
-    out = list(transform.apply(stream))
-    assert out == []
-
-
-def test_vector_drop_horizontal_with_only_and_exclude():
-    stream = iter(
-        [
-            _make_vector(0, {"a": 1.0, "b": None, "c": 2.0}),
-        ]
-    )
-    # Baseline features: a,b,c; with only=[a,b] and exclude=[b] -> baseline=[a]
-    # Coverage = coverage(a)=1.0, so vector kept.
-    transform = VectorDropTransform(
-        axis="horizontal",
-        threshold=1.0,
-        payload="features",
-        only=["a", "b"],
-        exclude=["b"],
-    )
-    transform.bind_context(_StubVectorContext(["a", "b", "c"]))
-
-    out = list(transform.apply(stream))
-    assert len(out) == 1
-
-
-def test_vector_drop_vertical_delegates_to_partitions():
-    stream = iter(
-        [
-            _make_vector(
-                0,
-                {
-                    "humidity__@location:north": 1.0,
-                    "humidity__@location:south": 2.0,
-                    "temp": 3.0,
-                },
-            )
-        ]
-    )
-    ctx = _StubVectorContext(
-        ["humidity__@location:north", "humidity__@location:south", "temp"]
-    )
-    ctx.set_metadata(
-        {
-            "counts": {"feature_vectors": 4},
-            "features": [
-                {"id": "humidity__@location:north", "present_count": 4, "null_count": 3},
-                {"id": "humidity__@location:south", "present_count": 4, "null_count": 0},
-                {"id": "temp", "present_count": 4, "null_count": 0},
-            ],
-        }
-    )
-    transform = VectorDropTransform(axis="vertical", threshold=0.8)
-    transform.bind_context(ctx)
-
-    out = list(transform.apply(stream))
-    assert list(out[0].features.values.keys()) == [
-        "humidity__@location:south",
-        "temp",
-    ]
-
-
-def test_vector_drop_partitions_drops_when_coverage_low():
-    stream = iter(
-        [
-            _make_vector(
-                0,
-                {
-                    "humidity__@location:north": 1.0,
-                    "humidity__@location:south": 2.0,
-                    "temp": 3.0,
-                },
-            )
-        ]
-    )
-    ctx = _StubVectorContext(
-        ["humidity__@location:north", "humidity__@location:south", "temp"]
-    )
-    ctx.set_metadata(
-        {
-            "counts": {"feature_vectors": 4},
-            "features": [
-                {"id": "humidity__@location:north", "present_count": 4, "null_count": 3},
-                {"id": "humidity__@location:south", "present_count": 4, "null_count": 0},
-                {"id": "temp", "present_count": 4, "null_count": 0},
-            ],
-        }
-    )
-    transform = VectorDropTransform(axis="vertical", threshold=0.8)
-    transform.bind_context(ctx)
-
-    ctx.load_schema()
-    out = list(transform.apply(stream))
-    assert list(out[0].features.values.keys()) == [
-        "humidity__@location:south",
-        "temp",
-    ]
-    schema = ctx._cache.get("schema:features")
-    if schema is None:
-        ctx.load_schema()
-        schema = ctx._cache["schema:features"]
-    assert [entry["id"] for entry in schema] == [
-        "humidity__@location:south",
-        "temp",
-    ]
-
-
-def test_vector_drop_partitions_keeps_when_coverage_sufficient():
-    stream = iter(
-        [
-            _make_vector(
-                0,
-                {
-                    "precip__@station:north": None,
-                    "precip__@station:south": 2.0,
-                    "temp": 10.0,
-                },
-            )
-        ]
-    )
-    ctx = _StubVectorContext(
-        ["precip__@station:north", "precip__@station:south", "temp"]
-    )
-    ctx.set_metadata(
-        {
-            "counts": {"feature_vectors": 4},
-            "features": [
-                {
-                    "id": "precip__@station:north",
-                    "present_count": 4,
-                    "null_count": 1,
-                },
-                {
-                    "id": "precip__@station:south",
-                    "present_count": 4,
-                    "null_count": 0,
-                },
-                {"id": "temp", "present_count": 4, "null_count": 0},
-            ],
-        }
-    )
-    transform = VectorDropTransform(axis="vertical", threshold=0.7)
-    transform.bind_context(ctx)
-
-    ctx.load_schema()
-    out = list(transform.apply(stream))
-    assert list(out[0].features.values.keys()) == [
-        "precip__@station:north",
-        "precip__@station:south",
-        "temp",
-    ]
-    schema = ctx._cache.get("schema:features")
-    if schema is None:
-        ctx.load_schema()
-        schema = ctx._cache["schema:features"]
-    assert [entry["id"] for entry in schema] == [
-        "precip__@station:north",
-        "precip__@station:south",
-        "temp",
-    ]
-
-
-def test_vector_drop_vertical_prefers_window_size_for_coverage():
-    stream = iter(
-        [
-            _make_vector(
-                0,
-                {
-                    "humidity__@location:north": 1.0,
-                },
-            )
-        ]
-    )
-    ctx = _StubVectorContext(["humidity__@location:north"])
-    ctx.set_metadata(
-        {
-            "window": {"start": "2024-01-01T00:00:00Z", "end": "2024-01-01T04:00:00Z", "size": 5},
-            "counts": {"feature_vectors": 100},  # would force coverage to near zero if used
-            "features": [
-                {
-                    "id": "humidity__@location:north",
-                    "present_count": 3,
-                    "null_count": 0,
-                },
-            ],
-        }
-    )
-    transform = VectorDropTransform(axis="vertical", threshold=0.5)
-    transform.bind_context(ctx)
-
-    out = list(transform.apply(stream))
-    assert out[0].features is not None
-    assert "humidity__@location:north" in out[0].features.values
-
-
-def test_vector_drop_vertical_uses_element_coverage_for_sequences():
-    stream = iter(
-        [
-            Sample(
-                key=(0,),
-                features=Vector(
-                    values={"pressure__@station_id:06183": [1.0, 2.0, 3.0, None]}
-                ),
-            )
-        ]
-    )
-    ctx = _StubVectorContext(["pressure__@station_id:06183"])
-    ctx.set_metadata(
-        {
-            "window": {"start": "2024-01-01T00:00:00Z", "end": "2024-01-01T01:00:00Z", "size": 1},
-            "counts": {"feature_vectors": 1},
-            "features": [
-                {
-                    "id": "pressure__@station_id:06183",
-                    "present_count": 1,
-                    "null_count": 0,
-                    "cadence": {"target": 4},
-                    "observed_elements": 3,  # one missing
-                }
-            ],
-        }
-    )
-    transform = VectorDropTransform(axis="vertical", threshold=1.0)
-    transform.bind_context(ctx)
-
-    out = list(transform.apply(stream))
-    assert out and out[0].features is not None
-    # Only 3/4 elements present => coverage < 1 => dropped
-    assert "pressure__@station_id:06183" not in out[0].features.values
-
-
-def test_vector_drop_partitions_accounts_for_absence():
-    stream = iter(
-        [
-            _make_vector(
-                0,
-                {
-                    "wind_speed__@station_id:06052": 1.0,
-                },
-            )
-        ]
-    )
-    ctx = _StubVectorContext(
-        ["wind_speed__@station_id:06052", "wind_speed__@station_id:06053"]
-    )
-    ctx.set_metadata(
-        {
-            "counts": {"feature_vectors": 4},
-            "features": [
-                {
-                    "id": "wind_speed__@station_id:06052",
-                    "present_count": 2,
-                    "null_count": 0,
-                },
-                {
-                    "id": "wind_speed__@station_id:06053",
-                    "present_count": 4,
-                    "null_count": 0,
-                },
-            ],
-        }
-    )
-    transform = VectorDropTransform(axis="vertical", threshold=0.75)
-    transform.bind_context(ctx)
-
-    out = list(transform.apply(stream))
-    assert list(out[0].features.values.keys()) == []
-
-
-def test_vector_drop_partitions_errors_without_metadata():
-    stream = iter([_make_vector(0, {"temp": 1.0})])
-    ctx = _StubVectorContext(["temp"])
-    transform = VectorDropTransform(axis="vertical", threshold=0.5)
-    transform.bind_context(ctx)
-
-    with pytest.raises(RuntimeError):
-        list(transform.apply(stream))
-
-
-def test_vector_drop_vertical_drops_null_target_at_threshold_one():
-    stream = iter(
-        [
-            Sample(
-                key=(0,),
-                features=Vector(values={}),
-                targets=Vector(
-                    values={
-                        "wind_production__@municipality_no:101": None,
-                        "wind_production__@municipality_no:147": 1.0,
-                    }
-                ),
-            )
-        ]
-    )
-    ctx = _StubVectorContext({"targets": [
-        "wind_production__@municipality_no:101",
-        "wind_production__@municipality_no:147",
-    ]})
-    ctx.set_metadata(
-        {
-            "counts": {"target_vectors": 1},
-            "targets": [
-                {
-                    "id": "wind_production__@municipality_no:101",
-                    "present_count": 1,
-                    "null_count": 1,
-                },
-                {
-                    "id": "wind_production__@municipality_no:147",
-                    "present_count": 1,
-                    "null_count": 0,
-                },
-            ],
-        }
-    )
-    transform = VectorDropTransform(
-        axis="vertical",
-        payload="targets",
-        threshold=1.0,
-    )
-    transform.bind_context(ctx)
-
-    out = list(transform.apply(stream))
-    assert len(out) == 1
-    assert out[0].targets is not None
-    assert set(out[0].targets.values.keys()) == {
-        "wind_production__@municipality_no:147"
-    }
-
-
-def test_vector_ensure_schema_errors_on_missing_by_default():
-    stream = iter([_make_vector(0, {"wind__A": 1.0})])
-    transform = VectorEnsureSchemaTransform()
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "wind__A"}, {"id": "wind__B"}]},
-        )
-    )
-
-    with pytest.raises(ValueError):
-        list(transform.apply(stream))
-
-
-def test_vector_ensure_schema_fill_and_reorder():
-    stream = iter([_make_vector(0, {"wind__B": 5.0, "wind__A": 2.0, "new": 9.0})])
-    transform = VectorEnsureSchemaTransform(
-        on_missing="fill", fill_value=0.0, on_extra="drop"
-    )
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "wind__A"}, {"id": "wind__B"}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert out[0].features.values == {"wind__A": 2.0, "wind__B": 5.0}
-
-
-def test_vector_ensure_schema_inserts_none_for_missing_slots():
-    stream = iter([_make_vector(0, {"wind__B": 5.0})])
-    transform = VectorEnsureSchemaTransform(on_missing="fill")
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "wind__A"}, {"id": "wind__B"}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert out[0].features.values == {"wind__A": None, "wind__B": 5.0}
-
-
-def test_vector_ensure_schema_orders_longer_baseline():
-    stream = iter([_make_vector(0, {"wind__C": 7.0, "wind__A": 1.0, "wind__B": 2.0})])
-    transform = VectorEnsureSchemaTransform()
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={
-                "features": [
-                    {"id": "wind__A"},
-                    {"id": "wind__B"},
-                    {"id": "wind__C"},
-                ]
-            },
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert [*out[0].features.values.keys()] == ["wind__A", "wind__B", "wind__C"]
-
-
-def test_vector_ensure_schema_drop_sample_on_missing():
-    stream = iter([_make_vector(0, {"wind__A": 2.0})])
-    transform = VectorEnsureSchemaTransform(on_missing="drop")
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "wind__A"}, {"id": "wind__B"}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert out == []
-
-
-def test_vector_ensure_schema_keep_extra_when_allowed():
-    stream = iter([_make_vector(0, {"wind__A": 1.0, "wind__B": 2.0, "wind__C": 3.0})])
-    transform = VectorEnsureSchemaTransform(on_extra="keep")
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "wind__A"}, {"id": "wind__B"}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert list(out[0].features.values.keys()) == ["wind__A", "wind__B", "wind__C"]
-
-
-def test_vector_ensure_schema_prefers_schema_artifact_over_expected_ids():
-    stream = iter([_make_vector(0, {"b": 2.0, "a": 1.0})])
-    transform = VectorEnsureSchemaTransform()
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "a"}, {"id": "b"}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert list(out[0].features.values.keys()) == ["a", "b"]
-
-
-def test_vector_ensure_schema_enforces_list_length_via_schema():
-    stream = iter([_make_vector(0, {"seq": [1, 2, 3]})])
-    transform = VectorEnsureSchemaTransform(on_missing="fill", fill_value=0)
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "seq", "kind": "list", "cadence": {"target": 5}}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert out[0].features.values["seq"] == [1, 2, 3, 0, 0]
-
-
-def test_vector_ensure_schema_enforces_length_error_on_violation():
-    stream = iter([_make_vector(0, {"seq": [1, 2, 3, 4]})])
-    transform = VectorEnsureSchemaTransform(on_missing="error")
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "seq", "kind": "list", "cadence": {"target": 2}}]},
-        )
-    )
-
-    with pytest.raises(ValueError):
-        list(transform.apply(stream))
-
-
-def test_vector_ensure_schema_respects_cadence_target_from_schema():
-    stream = iter([_make_vector(0, {"seq": [1, 2, 3, 4, 5]})])
-    transform = VectorEnsureSchemaTransform(on_missing="fill", fill_value=0)
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": [{"id": "seq", "kind": "list", "cadence": {"target": 6}}]},
-        )
-    )
-
-    out = list(transform.apply(stream))
-    assert len(out[0].features.values["seq"]) == 6
-
-
-def test_vector_ensure_schema_raises_without_schema_artifact():
-    stream = iter([_make_vector(0, {"wind__A": 1.0})])
-    transform = VectorEnsureSchemaTransform()
-    transform.bind_context(
-        _StubVectorContext(
-            [],
-            schema={"features": []},
-        )
-    )
-
-    with pytest.raises(RuntimeError):
-        list(transform.apply(stream))
-
-
-def test_vector_drop_null_feature_level_drops_when_coverage_low():
-    stream = iter([_make_vector(0, {"a": 1.0, "b": None})])
-    # a and b are expected; coverage = average(cell_coverage(a), cell_coverage(b))
-    # = (1.0 + 0.0)/2 = 0.5 < 0.8 -> drop
-    transform = VectorDropTransform(axis="horizontal", threshold=0.8)
-    transform.bind_context(_StubVectorContext(["a", "b"]))
-
-    out = list(transform.apply(stream))
-    assert out == []
-
-
-def test_vector_drop_null_feature_level_keeps_when_coverage_sufficient():
-    stream = iter([_make_vector(0, {"a": 1.0, "b": 2.0, "c": None})])
-    # expected: a,b,c -> coverage = (1.0 + 1.0 + 0.0)/3 ≈ 0.666 >= 0.5 -> keep
-    transform = VectorDropTransform(axis="horizontal", threshold=0.5)
-    transform.bind_context(_StubVectorContext(["a", "b", "c"]))
-
-    out = list(transform.apply(stream))
-    assert len(out) == 1
-
-
-def test_vector_drop_null_record_level_drops_on_sparse_sequence():
-    stream = iter([_make_vector(0, {"seq": [1.0, None, None], "f": 2.0})])
-    # seq coverage = 1/3, f coverage = 1.0 -> horizontal ≈ 0.666 < 0.8 -> drop
-    transform = VectorDropTransform(axis="horizontal", threshold=0.8)
-    transform.bind_context(_StubVectorContext(["seq", "f"]))
-
-    out = list(transform.apply(stream))
-    assert out == []
-
-
-def test_vector_drop_null_record_level_ignores_when_no_sequences():
-    stream = iter([_make_vector(0, {"a": 1.0})])
-    transform = VectorDropTransform(axis="horizontal", threshold=0.8)
-    transform.bind_context(_StubVectorContext(["a"]))
-
-    out = list(transform.apply(stream))
-    assert len(out) == 1
