+++ tests/unit/cli/test_contract_cli.py
