+++ tests/unit/cli/test_scaffold_plugin.py
