+++ tests/unit/cli/test_scaffold_source.py
