+++ tests/unit/cli/test_visuals_runner.py
