+++ tests/unit/io/test_output_resolution.py
