+++ tests/unit/pipeline/test_split_stage.py
